+++ src/extensions/workspace/ui/side-bar/component-tree/component-view/component-view.tsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useContext, useCallback } from 'react';
 import classNames from 'classnames';
 import { NavLink } from 'react-router-dom';
 import { TreeNodeProps } from '../recursive-tree';
@@ -9,35 +9,26 @@ import { clickable } from '../../../../../../to-eject/css-components/clickable';
 import { hoverable } from '../../../../../../to-eject/css-components/hoverable';
 import styles from './component-view.module.scss';
 
-export class ComponentView extends Component<TreeNodeProps> {
-  static contextType = ComponentTreeContext;
-  // context!: IComponentTreeContext;
+export function ComponentView(props: TreeNodeProps) {
+  const { node } = props;
+  const { onSelect } = useContext(ComponentTreeContext);
 
-  // handleClick = () => {
-  //   const { node } = this.props;
-  //   this.context.onSelect(node.id);
-  // };
+  const handleClick = useCallback(
+    (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
+      onSelect && onSelect(node.id, event);
+    },
+    [onSelect, node.id]
+  );
 
-  // private get isSelected() {
-  //   const { node } = this.props;
-  //   const { selected } = this.context;
-
-  //   return node.id === selected;
-  // }
-
-  render() {
-    const { node } = this.props;
-    // const {status} = this.contextType;
-
-    return (
-      <NavLink
-        to={`/${node.id}`}
-        className={classNames(indentClass, clickable, hoverable, styles.component)}
-        activeClassName={styles.active}
-      >
-        {getName(node.id)}
-        {/* <span>M</span> */}
-      </NavLink>
-    );
-  }
+  return (
+    <NavLink
+      to={`/${node.id}`}
+      className={classNames(indentClass, clickable, hoverable, styles.component)}
+      activeClassName={styles.active}
+      onClick={handleClick}
+    >
+      {getName(node.id)}
+      {/* <span>M</span> */}
+    </NavLink>
+  );
 }
